# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bot Hub API",
    "version": "1.0.0",
    "description": "Public REST API for skills. Rate limits: read 120/min per IP + 600/min per key; write 30/min per IP + 120/min per key."
  },
  "servers": [{ "url": "https://api.hanzo.ai" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API token"
      }
    },
    "schemas": {
...
      "TagMap": {
        "type": "object",
        "additionalProperties": { "type": "string" }
      },
      "Stats": {
        "type": "object",
        "additionalProperties": true
      },
      "SkillListItem": {
        "type": "object",
        "properties": {
          "slug": { "type": "string" },
          "displayName": { "type": "string" },
          "summary": { "type": ["string", "null"] },
          "tags": { "$ref": "#/components/schemas/TagMap" },
          "stats": { "$ref": "#/components/schemas/Stats" },
          "createdAt": { "type": "number" },
          "updatedAt": { "type": "number" },
          "latestVersion": {
            "type": ["object", "null"],
            "properties": {
              "version": { "type": "string" },
              "createdAt": { "type": "number" },
              "changelog": { "type": "string" }
            }
          }
        }
      },
      "SkillListResponse": {
        "type": "object",
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/SkillListItem" } },
          "nextCursor": { "type": ["string", "null"] }
        }
      },
      "Owner": {
        "type": ["object", "null"],
        "properties": {
          "handle": { "type": ["string", "null"] },
          "displayName": { "type": ["string", "null"] },
          "image": { "type": ["string", "null"] }
        }
      },
      "Skill": {
        "type": ["object", "null"],
        "properties": {
          "slug": { "type": "string" },
          "displayName": { "type": "string" },
          "summary": { "type": ["string", "null"] },
          "tags": { "$ref": "#/components/schemas/TagMap" },
          "stats": { "$ref": "#/components/schemas/Stats" },
          "createdAt": { "type": "number" },
          "updatedAt": { "type": "number" }
        }
      },
      "SkillResponse": {
        "type": "object",
        "properties": {
          "skill": { "$ref": "#/components/schemas/Skill" },
          "latestVersion": {
            "type": ["object", "null"],
            "properties": {
              "version": { "type": "string" },
              "createdAt": { "type": "number" },
              "changelog": { "type": "string" }
            }
          },
          "owner": { "$ref": "#/components/schemas/Owner" }
        }
      },
      "SkillVersion": {
        "type": "object",
        "properties": {
          "version": { "type": "string" },
          "createdAt": { "type": "number" },
          "changelog": { "type": "string" },
          "changelogSource": { "type": ["string", "null"], "enum": ["auto", "user", null] }
        }
      },
      "SkillVersionListResponse": {
        "type": "object",
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/SkillVersion" } },
          "nextCursor": { "type": ["string", "null"] }
        }
      },
...
      "PublishResponse": {
        "type": "object",
        "properties": {
          "ok": { "type": "boolean" },
          "skillId": { "type": "string" },
          "versionId": { "type": "string" }
        }
      },
      "DeleteResponse": {
        "type": "object",
        "properties": { "ok": { "type": "boolean" } }
      },
...
    }
  },
  "paths": {
...
    "/v1/skills": {
      "get": {
        "summary": "List skills",
        "parameters": [
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer" } },
          { "name": "cursor", "in": "query", "required": false, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Skills", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SkillListResponse" } } } }
        }
      },
      "post": {
        "summary": "Publish a skill version",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "payload": { "type": "string" },
                  "files": { "type": "array", "items": { "type": "string", "format": "binary" } }
                },
                "required": ["payload", "files"]
              }
            },
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "slug": { "type": "string" },
                  "displayName": { "type": "string" },
                  "version": { "type": "string" },
                  "changelog": { "type": "string" },
                  "tags": { "type": "array", "items": { "type": "string" } },
                  "forkOf": {
                    "type": "object",
                    "properties": { "slug": { "type": "string" }, "version": { "type": "string" } }
                  },
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "path": { "type": "string" },
                        "size": { "type": "number" },
                        "storageId": { "type": "string" },
                        "sha256": { "type": "string" },
                        "contentType": { "type": "string" }
                      }
                    }
                  }
                },
                "required": ["slug", "displayName", "version", "changelog", "files"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Published", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PublishResponse" } } } }
        }
      }
    },
    "/v1/skills/{slug}": {
      "get": {
        "summary": "Get skill",
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Skill", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SkillResponse" } } } },
          "404": { "description": "Not found" }
        }
      },
      "delete": {
        "summary": "Soft delete skill",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Deleted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DeleteResponse" } } } }
        }
      }
    },
    "/v1/skills/{slug}/undelete": {
      "post": {
        "summary": "Undelete skill",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Undeleted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DeleteResponse" } } } }
        }
      }
    },
    "/v1/skills/{slug}/versions": {
      "get": {
        "summary": "List versions",
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer" } },
          { "name": "cursor", "in": "query", "required": false, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Versions", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SkillVersionListResponse" } } } }
        }
      }
    },
...
  }
}